HS-241 | Labels refactor
